Optimize the hell out of it
Instead of doing a graph thingy with every study, studies are grouped by layer in the tree. There's one result per "powerset" of each layer instead of one result per element of said "powerset". Rather than revisiting nodes a million times and getting the same results in a different order, everything is linear. The only problem with this is that things like 302 have a bit less flexibility with when they can be bought, but there is still a very healthy spread of results for every TT value.